requirements.txt
monai
jupyter
nibabel
wandb
pytorch-ignite
scikit-image
mlflow
einops

hydra-core
hydra-colorlog
omegaconf